client/requirements.txt
paramiko>=3.0.0
pyyaml>=6.0
cryptography>=41.0.0
